Добавлено управление созданием продукции + UI Изменения

+++ TastyEat.Workstation/Models/Analytics/ProductionStatsDto.cs
@@ -4,7 +4,7 @@ public sealed class ProductionStatsDto
 {
     public string ProductName { get; set; } = string.Empty;
     public string ProductTypeName { get; set; } = string.Empty;
-    public int TotalProduced { get; set; }
-    public int TotalSold { get; set; }
-    public int Remaining { get; set; }
+    public double TotalProduced { get; set; }
+    public double TotalSold { get; set; }
+    public double Remaining { get; set; }
 }
